Made ints in arch_type_info 8 bit (because nothing reasonable will exceed 255)

+++ frontend/include/frontend/arch_type_info.h
@@ -2,21 +2,21 @@
 #define ARCH_TYPE_INFO_H
 
 #include <stdbool.h>
-#include <stddef.h>
+#include <stdint.h>
 
 #include "util/arch.h"
 
 struct arch_int_info {
-    size_t sint_size;
-    size_t int_size;
-    size_t lint_size;
-    size_t llint_size;
+    uint8_t sint_size;
+    uint8_t int_size;
+    uint8_t lint_size;
+    uint8_t llint_size;
 };
 
 struct arch_float_info {
-    size_t float_size;
-    size_t double_size;
-    size_t ldouble_size;
+    uint8_t float_size;
+    uint8_t double_size;
+    uint8_t ldouble_size;
 };
 
 struct arch_type_info {
